Users can now upload avatars via rails! This fixes #55.
The implementation is a bit tricky. Under the new rails system, images are uploaded to
secrets/uploads/user/avatar/WCA_ID/TIMESTAMP.EXT, and the filename is stored in
the users table in the avatar column. However, there are lots of people who
have uploaded profile images who do not yet have WCA accounts. To deal with
those people, we still need to hit the filesystem and check for them. If
people want to remove their avatar, they'll have to sign up for an account
and remove it through the edit profile page, which will set their avatar
column to null.