Do not escape rss content wrapped in CDATA.
This fixes #145.

Updating the specs was rather painful since oga is aware of CDATA and
only presents the inner part of the CDATA when calling `text` on a node.
So, in order to test whether the description content is actually wrapped
in CDATA I convert the node back to XML and test against that.
This is kinda fragile. Sorry.