Make Competitions.id immutable
Letting the end user change primary keys was a rather unfortunate design
decision which led to some weird corner cases. Especially keeping
foreign keys in sync was painful. This commit fixes that and introduces
another column: `friendly_id` which is meant to be shown and edited by the
user. `friendly_id` is generated by using the same rules as `id`.

Note: `id` is still generated before validation in order to not break
existing assumptions/code.