Cache competition results
I didn't cache the podiums and winners pages because caching them was
actually slower than not caching them at all (~65ms vs ~45ms). The
results/by_person and results/all pages are cached using a cache key
like:

    USNationals2015/2016-04-01T20:32:14Z/by_person/d9fb66b120b61f4...

The last part is generated by Rails and is the MD5 hash of the template
itself. Therefore it isn't technically necessary to include the string
"by_person", but I liked the explicity of it and I feel it's less
confusing for people not terrible familiar with fragment caching.

The second part is the `updated_at` column of the newest result for that
competition. There was no such column and adding one the Rails way
would've required us to update the PHP code, so I used a self-updating
column as described in:
http://gusiev.com/2009/04/update-and-create-timestamps-with-mysql/
As explained in the blog post you can't have two columns with `DEFAULT
NOW()` in MySQL without changing the `INSERT` code. That's why I didn't
add the `created_at` column as well.

The average response times for USNationals2015 dropped from ~700ms to
~150ms.