Add rubocop to enforce a consistent coding style
We spend quite some time correcting style issues in PRs, so I think we
could benefit from a tool which checks them for us. The actual rules
don't matter much as long as we're consistent. As a starting point
I've added multiline trailing commas as a rule since me and @jfly seem
to be in favor of it (helps preventing diff noise / merge conflicts).

Besides simply adding rubocop this also:

* Runs rubocop as part of travis
* Fixes some obvious rubocop violations like missing newline, trailing
  whitespaces, missing trailing commas, remove deprecated methods, ...
* Adds a `.rubocop_todo.yml` file which excludes all current violations.
  We should clean these up if we find time or simply disable that
  particular rule in `.rubocop.yml`.