Fix race condition in sign_up_spec.rb
When using the JS driver of Capybara the web server runs in a different
process than the specs, so writing expectations against the DB might
fail if the web server is not fast enough. Using an explicit

    expect(page).to have_content

forces Capybara to wait until the web server finished processing the
request.

This fixes #475.