Don't construct our own logger, because this just gets in the way when rails adds its own logger. See https://github.com/cubing/worldcubeassociation.org/issues/187#issuecomment-217261349 for some investigation.
This logger was added in 9e87d9d603da057cb6b2cbf7bec7d1bf5281d320, when
I probably was thinking that we'd run Unicorn in development mode.
Removing this might break things for people using Vagrant for
development (we might be using Unicorn in development mode there), but
the best solution for that is probably to change our Vagrant set up to
use `bin/rails s`. Or maybe we should just get rid of Vagrant and
dockerize all the things =)

This fixes #187.