Enforce extra validations when a competition is visible - Fixes #604
The default factory for a competition used to set showAtAll true without a delegate.
The extra validations make this default competition invalid. To change this the default
is now showAtAll false which matches what happens in the front end. A lot of tests needed
to be updated that created competitions because of this change.