Update Regulations history.
This involves a small change in directory structure and nginx configuration,
since this is the first time we have a previous Regulations version rendered by Rails.

Every history files not served by rails should now be in "files/" (which nginx handles), regs from 2015 and
upcoming updates should be placed in "official/".
(See the `regulations-data` branch for the layout)