Deleted registrations will be marked as such - fixes #1005
New columns deleted_at, and deleted_by have been added to the
registrations table to track when a registration is deleted.

A new column accepted_by has also been added to record the user id
of the person who accepted the registration.

A user can still only have 1 registration per competition. If the
registration is 'deleted' and then the user registers again, the
original registration will be updated and the deleted_at and accepted_at
columns will be set to null.