Refactor and comment the locale files.
Added context comment to the locale files: these comments will be displayed by the "Internationalize" app,
and will appear on a "key by key" basis, therefore the singular is used in these comments.

Commenting them showed that some keys could be reordered or grouped, so this has been done as well.