Set Stripe redirect_uri paramater and use WCA Test Stripe account.
When connecting an organizer's account to Stripe, a redirect_uri is
required so Stripe knows where to send it's response. Since we need
this code to work on a local machine and staging, we need to send this
redirect uri along with the connect request.

For our tests we also need a connected account so we can call the real
Stripe API. The competition factory now has an "entry_fee" trait which
will use this connected account.