Sort registrations by payment date.
When a competition does not have registration fees, the registration date is its creation date.
When a competition has registration fees, the registration is usually considered "full" when payment is received.

The sort acts as follow :
 - Paid registration are displayed first, and sorted by last payment date
 - Unpaid registrations are then sorted by creation date

This keeps the current behavior for competitions not using payment with Stripe, and enable sorting by payment date for competitions using Stripe.