Move test around
Using javascript on the "show" page raises a js error that occurs only in capybara's browser, where "Intl" is not defined.
(which is used by wca_local_time)

We can actually test the javascript feature on the index page (where we don't have a reference to Intl), so we do this instead.