Add privacy statement to the website.
There's a link to /privacy in our footer, and in our navbar.

This was a fairly manual process that started with downloading a Google
Doc as html, and then manually cleaning it up. I documented all the
steps I took in a comment at the top of privacy.html.erb.