Don't unconditionally compute derived state.
Before this change, resizing the page would cause any "draft"
state in our SelectRoundsButton to disappear! This is because
resizing would cause componentWillReceiveProps to be called, and that
would clobber our components draft state.

This article goes into some discussion about this class of bugs, along
with some suggestions for fixing it. I decided to go with the "Fully
uncontrolled component with a key" solution detailed here:
https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html#recommendation-fully-uncontrolled-component-with-a-key.
However, we don't actually need a key here because the only way to trigger a
change to the modal is to close the modal, so the fix is to just compute
our state from props once when the component is created, and to *not*
recompute it when our props change.