Bump dynamic HSTS max-age from half a year to 2 years and update associated comment. (#2953)
worldcubeassocation.org has been preloaded for a while, which means we're
committed HSTS for the long term. We also haven't had any issues with HSTS for
subdomains, so we might as well extend the dynamic max-age in case we have users
without preloaded HSTS who visit slightly less than once a year (e.g. for an
annual competition that isn't always near the same date in the year).