Remove column names from generated csv voter files (#3259)
* Remove column names from generated csv voter files

From Walker via email:
I just want to make a minor change to the voting files that are generated when I click on each one.

In each file, it has the first row as a list of: Name, email, ID. Would I be able to get that row removed? 

When I try to upload the file to Helios, it tries to add that row into the data as a voter. So each time I have to take it out of the data before uploading the file. Sometimes messing around with the file also makes it so the file doesn't upload correctly in Helios.