Actually show all fields to organizers/delegates.
Ron says "I propose to show all fields in read mode on the organizer's
screen. With the recent generation of registration requirements it has
become less clear which text was generated and which text was entered.
But even without this, it is always handy to have access to the
originally entered information."

I totally agree with this sentiment.