Add a CityValidator for validating city names.
(This is part of #361.)

I had some fun with Ruby here. My thinking is that the validations we
want to enforce will be a bit different depending on the country, so I
made it possible to create a `CountryCityValidator` for each
country. These `CountryCityValidator`s know how to check if a city name
is valid or not. If there is no `CountryCityValidator` for a given
country, then we treat the city name as valid. I could see us changing
that in the future once we've added a `CountryCityValidator` for every
country we host competitions in.