Add a `confirmed_at` column to Competitions and remove the `isConfirmed` column.
`confirmed_at` NULL means the competition is not confirmed.
`confirmed_at` not NULL means the competition is confirmed. When
migrating old, confirmed competitions, we set `confirmed_at` to
`announced_at` if `announced_at` is set, else we set `confirmed_at` to
`NOW()`.

The WQAC wants to handle incoming competition requests in the order they
were confirmed, and currently we don't have that information anywhere in
the database. This solves that problem. See
https://github.com/thewca/worldcubeassociation.org/issues/3321 for more
information.