Some cleanup of hidden teams.
- The association between TeamMember -> Team wasn't working because of
the default scope on Team which would prevent finding the team for
banned team members.
- A similar issue would prevent `Team.banned` from finding the cached
team, because we were doing a `Team.all` instead of a `Team.unscoped`.
- When dumping the developer database export, we actually *do* want to
include hidden teams, we just don't want to include the members of those
teams. Previously, we were excluding the teams, but not the team
members, which means that members of the banned team would actually end
up in the developer export! Fortunately, we haven't started using the
banned competitors team yet, so we haven't leaked any sensitive
information. Also, it's better to actually include the banned team in
the export because any code that does a `Team.banned` will break if the
banned competitors team does not exist.