Add a stripe_charges table to track unknown, successful, and failed payments.
We had a weird issue a while back where we successfully charged someone,
but we didn't record the payment. We didn't find out about this until
after our logs had rotated away. This change makes it possible to detect
this situation programatically (the presence of stripe_charges in
"unknown" status means that we tried to talk to Stipe, and we don't know
what happened after that).

I will follow this commit with a change to our server-status page to
actually detect this situation.