Add support for syncing leaders@ and results@
I imagine this will go well. If it does, then it should be
straightforward to implement the rest of the forwards from
https://docs.google.com/spreadsheets/d/1I2l_ht2NhcXurG9wezludKNb6ZFjpuWS4jrF_h8DaIw/edit?ts=5bef0e17#gid=0.

I'm not sure what the nicest way to write the tests for this is. The way
I've done it is one big test, which is a bit tricky to read through
(testing leaders@ interacts with testing results@, for instance). I'd
like to have one test per mailing list, but I couldn't figure out how to
`expect` one call to `GsuiteMailingLists.sync_group` but ignore the
other calls.