Add SyncMailingListsJob to the schedule work task.
This means it will run once an hour. We should probably also schedule
the job whenever something changes that would cause mailing list
membership to change, but I think it would still be good to periodically
run the job once an hour just in case we forget to schedule the job.