Allow sync-ing google groups with an email address rather than a user.
Email addresses are a little more flexible, because we don't
necessarily have a user in the database for every email address we may
want to add to a mailing list.

We actually wrote ede78479 assuming you *could* pass email addresses into
`sync_group`, but that's causing job crashes with errors like this:

```
Job 15999 failed. #<NoMethodError: undefined method `email' for "seniors@worldcubeassociation.org":String>
```

One of the downsides of using mocks to test things :(