Do not require entry fees for competitions in multiple countries.
This just came up for FMC Europe, a competition which has different
registration requirements per venue. Rather than requiring the user to
specify an inaccurate number, we'd rather have these values be NULL
(unknown) in our database.

Note: This does not detect competitions in multiple venues within the
same country (such as FMC USA). We can add support for that later if we
decide we want to.