Wrap announcing competition in a database transaction.
Jacob noticed that two competitions in our database were announced, but
their `announced_at` columns were NULL. This must have been because we
crashed somewhere inbetween the call to `create_post_and_redirect` and
the call to `comp.update!`. Maybe the competition failed validations for
some reason? It's hard to tell what happened, because our server logs don't go
back that far. If we had had this code wrapped in a transaction
previously, then announcing the competition would have failed, and we
would have known to investigate it.

Note, I also moved the sending of the email to happen outside of the
database transaction. This feels more correct to me, as we don't want to
email people unless we actually *did* announce the competition (by
creating a post and updating the `announced_at` column).