Store user's preferred locale when signing up.
This fixes #3082. Previously, when signing up, a user would see the website in their
preferred language (because their browser sends us the
`Accept-Lanaguage` header). Their `preferred_locale` would remain NULL
in the database. Then, when they sign back into the website, they'd
still see the website in their preferred language (because they don't
have a `preferred_locale` set, we'd fall back to their `Accept-Language`
header. So far, so good. However, when we'd go on to email the user, we
don't have any HTTP headers to look at to determine their preferred
language, so we'd email them in English, which is pretty silly, because
we should have known that they prefer a different language.

This change stores the locale we derive from a users's
`Accept-Lanaguage` header in their `preferred_locale` column, so emails
should end up in their preferred language.

One quirk of this change is that a user whose language is *not*
supported by the website will end up getting their `preferred_locale`
set to `en`. I don't think this is a problem, but may be a little suprising.