Stricter date parsing.
Our UI should only send dates of the form YYYY-MM-DD, so it would be
nice for our backend to complain if dates do not match that format.
The previous implementation would happily parse dates like `2018-1-3`,
which I think is better to not allow.

The Results Team just notified us of a dob change request with a date of
the form YYYY-M-DD, which means something must have gone wrong with our
datepicker. This doesn't fix that issue, but maybe changes things so
it's more likely people will contact us if they have issues with the
datepicker, and we can track that down.