Revert "Stricter date parsing."
This reverts commit b4607ccf041ac792572190980f479d9e59051c63.

Jonatan points out that lots of clients (such as TNoodle
https://github.com/thewca/tnoodle/blob/8152b1dd2a07647aa5287506c3551abff09973de/tnoodle-ui/src/WcaApi.js#L60)
actually do send us datetime strings, and we don't want to break that.

I'll merge this up immediately to avoid breaking anything, and then
we can follow up with a better fix, because I would like us to be string
about 2 character months and dates.