Some more changes to stripe charge handling:
We want journaled_stripe_charge to *also* handle
`Stripe::InvalidRequestError` so we can mark the charge as failed.

Also DRY up the stripe exception handling one layer higher in
`process_payment`. We had forgotten to add a `redirect_to` in the
`Stripe::InvalidRequestError` case, and I think it's fine to handle them
in the same rescue block for now (someday we may want to split them up
again if we want to handle them differently).