Fix error with missing rounds
When submitting results we may enter a situation where one of the round
present in a cumulative time limit doesn't exist in the actual results.
In this case the error is already reported as a missing round, but then
we should not assume it exists when checking individual results.

It also fixes an issue when the WCIF gives a cumulative time limit with
a round which does not exist in the same WCIF. It should not happen but
it may because of bug #3254.