Add a proxy API endpoint for geocoding
We want to restrict our Google API key on an IP basis, so we need to
make sure the actual query comes from our server (and not from the
client).
We also want to make sure anyone on the internet can't just use our
proxy, so we explicitly check for Rails' CSRF token even if we're just
doing a GET query (this way we are sure the query made by the client is
legit).