Do not allow trailing commas in city names.
Ruby's split method will split as many times as possible, and when
destructuring the result into 2 variables, it only pulls off the first
two pieces, and drops the rest. This is not what we want, because it
allows people to (intentionally or accidentally) add more stuff to city
names after commas and we won't realize it.

I first added a failing test that passes now that we limit the split to
producing 2 items.