Add support for uploading images in markdown fields!
Support for uploading images was very recently added to EasyMDE.
I followed
https://medium.com/weareevermore/manual-uploads-using-activestorage-47808dab1b65
to build a new endpoint just for uploading images, and then wrote a
little bit of javascript to actually connect EasyMDE to this new
endpoint. Unfortunately, I couldn't find a good way to add validations
to this new endpoint. Since we only allow fairly trusted users to access
this endpoint, I think that's acceptable for now, but eventually it
would be nice to validate file size/format/etc. See
https://github.com/thewca/worldcubeassociation.org/issues/4380 for more
information.

This is part of https://github.com/thewca/worldcubeassociation.org/issues/4309.

I found a problem with EasyMDE's image upload button: it doesn't seem to
use the custom image upload function if you've specified one. It was
fairly easy to work around this issue, but I have filed
https://github.com/Ionaru/easy-markdown-editor/pull/109 upstream to fix
the bug in the right place.