Switch from `rails-sass` to `rails-sassc`
This is a step towards https://github.com/thewca/worldcubeassociation.org/issues/3649 =)

As far as I can tell, this is a drop in replacement. In fact, it looks
like `rails-sass` is already a meaningless shell around `rails-sassc`:
https://github.com/rails/sass-rails/pull/424, so I think this is
basically a no-op. Regardless, it feels cleaner to me to use
`rails-sassc` directly.