Adds current version of the Regulation to the history index
This commit adds the current version of the Regulations to the history page of the Regulations.

A WCA Delegate recently approached the WRC and pointed out that even though the history page of the Regulations states that the page shows all past and current versions of the Regulations, only the past versions were being displayed.

This commit therefore adds the current (2020-08-01) version of the Regulations to the history page.

It makes it easier for someone to access what was changed in the current version of the Regulations.

Besides, after some research, it seems like this was what was practiced in the past (as in [this](https://github.com/thewca/worldcubeassociation.org/commit/f088abb0da19253d9b6972f7747bbcdf1afcb737#diff-dfdcdad65271fa74be24e90039016542) commit), but was forgotten in the 2019 Regulations update.