Add option to require qualifications.
Adds an option to the competition form to require that competitors be qualified for events that they are registering for.

This does not prevent registering for events with a "Top N" qualification, because it's not yet known what the qualifying standard will be. In theory, we could do "Top N of people currently registered", but that would lead to weird effects if someone qualified cancelled their registration.

Part of #2051.