Add option to require qualification for registered events. (#7019)
* Update qualifications to new WCIF spec.

* Add option to require qualifications.

Adds an option to the competition form to require that competitors be qualified for events that they are registering for.

This does not prevent registering for events with a "Top N" qualification, because it's not yet known what the qualifying standard will be. In theory, we could do "Top N of people currently registered", but that would lead to weird effects if someone qualified cancelled their registration.

Part of #2051.

* Fix lint errors.

* Fix tests.

* Move the qualification options together.

* People without WCA ID aren't qualified.

* Code review comments.

* Some more changes from code review.

* Fix tests.

* Use & instead of explicit nil check.

* Code review comments.

* Fix formatting issues.

* Code review comments.

* Fix extraneous merge markings.

* Remove generated files from app/webpacker/packs/generated_locales.

* Fix structure.sql. Again.

* More structure.sql fixes.

* Move error message to the right place.

Had a bad merge at some point.