Include non-competing staff in the public WCIF export (#8001)
* Include non-competing staff in the public WCIF export

* Use the same WCIF registration status for non-competing

Instead of defining a new status, use the same one as accepted
competitors.

* Add isCompeting field to Registration wcif

Added to WCIF schema in thewca/wcif#15.

* Only add non-competing registrations via PATCH

This requires the app calling the WCIF PATCH to specify

registration: {
  is_competing: false
}

in order to add a person. This helps ensure that this is intentional
(i.e. don't just add all organizers automatically).

* Code review comments

* Code review fix