Improve ranking links in user profile
Links to the rankings page are put on the NR/CR/WR numbers instead, and redirect
to their respective region-based rankings (national, continental, world). When
one of these links is clicked, the "focus" parameter highlights the user's row
in the rankings if applicable and scrolls it to the center of the viewport.