Use ruby instead of sed in mysqldump
sed's -i option isn't a part of POSIX and gets implemented differently
on GNU and BSD. As a result the way it's used in
DatabaseDumper::mysqldump isn't cross-platform.

Since we can reasonably expect ruby to be installed on the target
platforms for this app, we can also use ruby instead of sed in this
instance.

See this discussion for further info:

https://stackoverflow.com/questions/5694228/sed-in-place-flag-that-works-both-on-mac-bsd-and-linux