Don't display team/committee acronym in header (#10692)
This commit removes the code that displays the team/committee `friendly_id`, which has additional meaning under-the-hood that was leading to incorrect display of acronyms. In particularly, Appeals Committee was being displayed as "WAPC" instead of the desired "WAC".

This is not a complete fix, as `#wapc` is still used as the URL fragment when viewing the Appeals Committee page, but it does resolve the more visible instance of the wrong acronym.

This commit also adjusts the langauge in the en locale file to use the defined acronym, and aligns with best technical writing practices when applicable (plus some other small improvements/changes to team responsibilities).